requirements.txt
Django==5.1.4
djangorestframework==3.15.2
Faker==33.1.0
xhtml2pdf==0.2.16